Implemented GREENLIGHT state, tested out new motors and improved calibAccel output.
GREENLIGHT state simply lights the LED up in green. For use while debugging flags.
New motors are good for balancing but crap for racing. Too slow.
The calibAccel.c code now outputs the mean X, Y, Z and Magnitude for the accelerometer.
It was found out that the accelerometer is biased. Working on it.